// shared/warframeRankExceptions.generated.ts
export const WARFRAME_RANK_EXCEPTIONS_BY_UNIQUE_NAME: Readonly<
  Record<string, RankStatBonuses>
> = {
  "/Lotus/Powersuits/Pacifist/Pacifist": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Pacifist/BaruukPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Dragon/ChromaPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Pagemaster/Pagemaster": {
    "health": 90,
    "shield": 90,
    "armor": 0,
    "energy": 70
  },
  "/Lotus/Powersuits/Garuda/Garuda": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Garuda/GarudaPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Devourer/Devourer": {
    "health": 200,
    "shield": 0,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Devourer/GrendelPrime": {
    "health": 200,
    "shield": 0,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/IronFrame/IronFrame": {
    "health": 100,
    "shield": 500,
    "armor": 0,
    "energy": 0
  },
  "/Lotus/Powersuits/IronFrame/IronFramePrime": {
    "health": 100,
    "shield": 500,
    "armor": 0,
    "energy": 0
  },
  "/Lotus/Powersuits/Sandman/Sandman": {
    "health": 200,
    "shield": 0,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Sandman/InarosPrime": {
    "health": 200,
    "shield": 0,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Koumei/Koumei": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/PaxDuviricus/PaxDuviricus": {
    "health": 200,
    "shield": 0,
    "armor": 100,
    "energy": 50
  },
  "/Lotus/Powersuits/Alchemist/Alchemist": {
    "health": 200,
    "shield": 100,
    "armor": 100,
    "energy": 0
  },
  "/Lotus/Powersuits/Alchemist/LavosPrime": {
    "health": 200,
    "shield": 100,
    "armor": 100,
    "energy": 0
  },
  "/Lotus/Powersuits/Nezha/Nezha": {
    "health": 100,
    "shield": 50,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Nezha/NezhaPrime": {
    "health": 100,
    "shield": 50,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Infestation/Infestation": {
    "health": 100,
    "shield": 0,
    "armor": 100,
    "energy": 50
  },
  "/Lotus/Powersuits/Infestation/InfestationPrime": {
    "health": 100,
    "shield": 0,
    "armor": 100,
    "energy": 50
  },
  "/Lotus/Powersuits/Saryn/SarynPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Berserker/Berserker": {
    "health": 100,
    "shield": 50,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Berserker/ValkyrPrime": {
    "health": 100,
    "shield": 50,
    "armor": 0,
    "energy": 50
  },
  "/Lotus/Powersuits/Volt/VoltPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Wisp/Wisp": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Wisp/WispPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/BrokenFrame/BrokenFrame": {
    "health": 90,
    "shield": 90,
    "armor": 0,
    "energy": 70
  },
  "/Lotus/Powersuits/BrokenFrame/XakuPrime": {
    "health": 90,
    "shield": 90,
    "armor": 0,
    "energy": 70
  },
  "/Lotus/Powersuits/Yareli/Yareli": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  },
  "/Lotus/Powersuits/Yareli/YareliPrime": {
    "health": 100,
    "shield": 100,
    "armor": 0,
    "energy": 100
  }
};